Move SolusParseException to the Exceptions folder.
